Rename docs/assets/buffer*.js
