+++ docs/assets/buffer_coordseq.js
